// src/utils/enums/serviceToken.ts
enum ServiceToken {
  AUTH_SERVICE = 'AuthService',
  USER_SERVICE = 'UserService',
  MAIL_SERVICE = 'MailService',
  MAIL_OPTION_SERVICE = 'MailOptionService',
  PLAN_SERVICE = 'PlanService',
  PLAN2_SERVICE = 'Plan2Service',
  REFERRAL_SETTINGS_SERVICE = 'referralSettingsService',
  REFERRAL_SERVICE = 'referralService',
  NOTIFICATION_SERVICE = 'notificationService',
  FAILED_TRANSACTION_SERVICE = 'failedTransactionService',
  TRANSACTION_MANAGER_SERVICE = 'transactionManagerService',
  EMAIL_VERIFICATION_SERVICE = 'emailVerificationService',
  RESET_PASSWORD_SERVICE = 'resetPasswordService',
  SEND_MAIL_SERVICE = 'SendMailService',
  ACTIVITY_SERVICE = 'activityService',
  CURRENCY_SERVICE = 'currencyService',
  DEPOSIT_METHOD_SERVICE = 'depositMethodService',
  WITHDRAWAL_METHOD_SERVICE = 'withdrawalMethodService',
  DEPOSIT_SERVICE = 'depositService',
  WITHDRAWAL_SERVICE = 'withdrawalService',
  TRANSACTION_SERVICE = 'transactionService',
  TRANSFER_SETTINGS_SERVICE = 'transferSettingsService',
  TRANSFER_SERVICE = 'transferService',
  ASSET_SERVICE = 'assetService',
  INVESTMENT_SERVICE = 'investmentService',
  ITEM_SETTINGS_SERVICE = 'itemSettingsService',
  ITEM_SERVICE = 'itemService',
  PAIR_SERVICE = 'pairService',
  TRADE_SERVICE = 'tradeService',
  FORECAST_SERVICE = 'forecastService',
  MATH_SERVICE = 'mathService',
  MATH_UTILITY = 'mathUtility',
}

export default ServiceToken